+++ mwparserfromhell/nodes/extras/attribute.py
@@ -36,12 +36,14 @@ class Attribute(StringMixIn):
     whose value is ``"foo"``.
     """
 
-    def __init__(self, name, value=None, quoted=True, pad_first=" ",
+    def __init__(self, name, value=None, quotes='"', pad_first=" ",
                  pad_before_eq="", pad_after_eq=""):
         super(Attribute, self).__init__()
+        if not quotes and self._value_needs_quotes(value):
+            raise ValueError("given value {0!r} requires quotes".format(value))
         self._name = name
         self._value = value
-        self._quoted = quoted
+        self._quotes = quotes
         self._pad_first = pad_first
         self._pad_before_eq = pad_before_eq
         self._pad_after_eq = pad_after_eq
@@ -50,10 +52,17 @@ class Attribute(StringMixIn):
         result = self.pad_first + str(self.name) + self.pad_before_eq
         if self.value is not None:
             result += "=" + self.pad_after_eq
-            if self.quoted:
-                return result + '"' + str(self.value) + '"'
+            if self.quotes:
+                return result + self.quotes + str(self.value) + self.quotes
             return result + str(self.value)
         return result
+
+    @staticmethod
+    def _value_needs_quotes(val):
+        """Return the preferred quotes for the given value, or None."""
+        if val and any(char.isspace() for char in val):
+            return ('"' in val and "'" in val) or ("'" if '"' in val else '"')
+        return None
 
     def _set_padding(self, attr, value):
         """Setter for the value of a padding attribute."""
@@ -64,6 +73,14 @@ class Attribute(StringMixIn):
             if not value.isspace():
                 raise ValueError("padding must be entirely whitespace")
             setattr(self, attr, value)
+
+    @staticmethod
+    def coerce_quotes(quotes):
+        """Coerce a quote type into an acceptable value, or raise an error."""
+        orig, quotes = quotes, str(quotes) if quotes else None
+        if quotes not in [None, '"', "'"]:
+            raise ValueError("{0!r} is not a valid quote type".format(orig))
+        return quotes
 
     @property
     def name(self):
@@ -76,9 +93,9 @@ class Attribute(StringMixIn):
         return self._value
 
     @property
-    def quoted(self):
-        """Whether the attribute's value is quoted with double quotes."""
-        return self._quoted
+    def quotes(self):
+        """How to enclose the attribute value. ``"``, ``'``, or ``None``."""
+        return self._quotes
 
     @property
     def pad_first(self):
@@ -101,11 +118,21 @@ class Attribute(StringMixIn):
 
     @value.setter
     def value(self, newval):
-        self._value = None if newval is None else parse_anything(newval)
+        if newval is None:
+            self._value = None
+        else:
+            code = parse_anything(newval)
+            quotes = self._value_needs_quotes(code)
+            if quotes in ['"', "'"] or (quotes is True and not self.quotes):
+                self._quotes = quotes
+            self._value = code
 
-    @quoted.setter
-    def quoted(self, value):
-        self._quoted = bool(value)
+    @quotes.setter
+    def quotes(self, value):
+        value = self.coerce_quotes(value)
+        if not value and self._value_needs_quotes(self.value):
+            raise ValueError("attribute value requires quotes")
+        self._quotes = value
 
     @pad_first.setter
     def pad_first(self, value):